+++ CHANGELOG.md
@@ -836,3 +836,7 @@ a
 
 ## 0.1.650
 * Fix on internal padding of ZwapFloatingPicker and forced placeholder to fit in 1 line
+
+## 0.1.651
+* Forced placeholder to fit in 1 line in ZwapCheckBoxPicker Widget
+* Added ZwapSearchPicker Widget